requirements.txt
phidata
PyGithub
python-dotenv
Iterator
openai
groq
gitpython
zipfile
streamlit
uvicorn 
fastapi